backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pillow
numpy
ultralytics
python-multipart
opencv-python-headless
